Rename sdl.h header to avoid conflict
This avoids a conflict with SDL's own SDL.h on
case-insensitive filesystems.
