+++ src/Makefile.am
@@ -3,7 +3,7 @@ bin_PROGRAMS = adplay
 adplay_SOURCES = adplay.cc output.cc output.h players.h defines.h
 
 EXTRA_adplay_SOURCES = oss.cc oss.h null.h disk.cc disk.h esound.cc esound.h \
-	qsa.cc qsa.h sdl.cc sdl.h alsa.cc alsa.h ao.cc ao.h getopt.c \
+	qsa.cc qsa.h sdl.cc sdl_driver.h alsa.cc alsa.h ao.cc ao.h getopt.c \
 	getopt1.c getopt.h
 
 adplay_LDADD = $(drivers) $(adplug_LIBS) @ESD_LIBS@ @QSA_LIBS@ @SDL_LIBS@ \
